Relocate IconScoreDatabase
* database.db -> iconscore.db
