+++ tests/legacy_unittest/database/test_db_db.py
@@ -23,12 +23,12 @@ from iconservice.base.address import Address, AddressPrefix
 from iconservice.base.exception import DatabaseException, InvalidParamsException
 from iconservice.database.batch import BlockBatch, TransactionBatch, TransactionBatchValue, BlockBatchValue
 from iconservice.database.db import ContextDatabase, MetaContextDatabase
-from iconservice.database.db import IconScoreDatabase
 from iconservice.database.db import KeyValueDatabase
+from iconservice.database.wal import StateWAL
 from iconservice.icon_constant import DATA_BYTE_ORDER
+from iconservice.iconscore.db import IconScoreDatabase
 from iconservice.iconscore.icon_score_context import IconScoreContextType, IconScoreContext
 from iconservice.iconscore.icon_score_context import IconScoreFuncType
-from iconservice.database.wal import StateWAL
 from tests import rmtree
 
 
@@ -289,7 +289,7 @@ class TestIconScoreDatabase(unittest.TestCase):
         db_path = os.path.join(state_db_root_path, 'db')
         context_db = ContextDatabase.from_path(db_path, True)
 
-        self.db = IconScoreDatabase(address, context_db=context_db, prefix=b'')
+        self.db = IconScoreDatabase(address, context_db=context_db)
         self.address = address
 
     def tearDown(self):
