+++ tests/unit_test/iconscore/test_external_payable_call_method.py
@@ -22,10 +22,10 @@ from iconservice.base.block import Block
 from iconservice.base.exception import ExceptionCode, MethodNotFoundException
 from iconservice.base.message import Message
 from iconservice.base.transaction import Transaction
-from iconservice.database.db import IconScoreDatabase
 from iconservice.deploy import DeployEngine
 from iconservice.icon_constant import IconScoreContextType, IconScoreFuncType
 from iconservice.iconscore.context.context import ContextContainer
+from iconservice.iconscore.db import IconScoreDatabase
 from iconservice.iconscore.icon_score_base import IconScoreBase, external, payable
 from iconservice.iconscore.icon_score_constant import ATTR_SCORE_CALL
 from iconservice.iconscore.icon_score_context import IconScoreContext
