+++ tests/unit_test/iconscore/test_icon_score_trace.py
@@ -18,7 +18,7 @@
 """
 import os
 from random import randrange
-from unittest.mock import Mock, patch
+from unittest.mock import Mock
 
 import pytest
 
@@ -28,11 +28,11 @@ from iconservice.base.exception import ExceptionCode, IconScoreException, Invali
 from iconservice.base.message import Message
 from iconservice.base.transaction import Transaction
 from iconservice.database.batch import TransactionBatch
-from iconservice.database.db import IconScoreDatabase
 from iconservice.deploy import DeployEngine, DeployStorage
 from iconservice.icon_constant import IconScoreContextType
 from iconservice.icon_service_engine import IconServiceEngine
 from iconservice.iconscore.context.context import ContextContainer
+from iconservice.iconscore.db import IconScoreDatabase
 from iconservice.iconscore.icon_pre_validator import IconPreValidator
 from iconservice.iconscore.icon_score_base import IconScoreBase, external, interface
 from iconservice.iconscore.icon_score_base2 import InterfaceScore
